requirements.txt
asgiref==3.8.1
defusedxml==0.7.1
Django==3.2.25
django-filter==23.3
djangorestframework==3.15.1
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
